+++ packages/polaris-viz/src/components/LineChartRelational/components/RelatedAreas/index.ts
@@ -1,0 +1,1 @@
+export {RelatedAreas} from './RelatedAreas';