+++ packages/polaris-viz/src/components/LineChartRelational/index.ts
@@ -1,0 +1,1 @@
+export {LineChartRelational} from './LineChartRelational';