+++ packages/polaris-viz/src/components/TooltipContent/components/TooltipContentContainer/TooltipContentContainer.tsx
@@ -1,14 +1,38 @@
-import {changeColorOpacity, useTheme} from '@shopify/polaris-viz-core';
+import {
+  COLOR_VISION_SINGLE_ITEM,
+  changeColorOpacity,
+  useTheme,
+} from '@shopify/polaris-viz-core';
+import type {ReactNode} from 'react';
+import {useState} from 'react';
 
+import {useWatchColorVisionEvents} from '../../../../hooks';
 import {TOOLTIP_BG_OPACITY} from '../../../../constants';
 import {useBrowserCheck} from '../../../../hooks/useBrowserCheck';
 
 import styles from './TooltipContentContainer.scss';
 
-export function TooltipContentContainer({children, maxWidth, theme}) {
+interface Props {
+  children: ({
+    activeColorVisionIndex,
+  }: {
+    activeColorVisionIndex: number;
+  }) => ReactNode;
+  maxWidth: number;
+  theme: string;
+}
+
+export function TooltipContentContainer({children, maxWidth, theme}: Props) {
   const {isFirefox} = useBrowserCheck();
 
   const selectedTheme = useTheme(theme);
+
+  const [activeColorVisionIndex, setActiveIndex] = useState(-1);
+
+  useWatchColorVisionEvents({
+    type: COLOR_VISION_SINGLE_ITEM,
+    onIndexChange: ({detail}) => setActiveIndex(detail.index),
+  });
 
   return (
     <div
@@ -21,7 +45,7 @@ export function TooltipContentContainer({children, maxWidth, theme}) {
         maxWidth,
       }}
     >
-      {children}
+      {children({activeColorVisionIndex})}
     </div>
   );
 }